remove composite span processor
